Add files via upload
This project was part of my second-semester coursework. Neighbors Green Grocer is a real business that has been part of my life since I was 14 years old, as my mother has been a customer there for the past 10 years. I chose to name my project after this place to create a sense of familiarity and to motivate myself to put in my best effort, as this was my final undergraduate project. I apologize in advance for any inconsistencies in the file arrangements, as I was learning and implementing the project simultaneously.